requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
Deprecated==1.2.18
fastapi==0.115.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyGithub==2.5.0
PyJWT==2.10.1
PyNaCl==1.5.0
pytest==8.3.4
python-dotenv==1.0.1
requests==2.32.3
slack_sdk==3.34.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
uvicorn==0.34.0
wrapt==1.17.2
